[김민정, 김서영] 로그인 session 처리, 자기소개서 리스트 출력, 수정, 삭제, 등록 구현 및 뷰 생성 1. 각 기능마다 session 값을 넣어서 session값이 없으면 기능을 구현하지 못하도록 하는 코드 추가 2. SelfIntroductionMapper.xml 	리스트 불러와 출력하는 쿼리 생성 	자소서 등록하는 쿼리 생성 	자조서 수정하기 위해 보여지는 폼 쿼리 생성 	자소서 수정하는 쿼리 생성 	자소서 삭제하는 쿼리 생성 3. SelfIntroductionMapper interface 	자소서 기능 구현하는 쿼리 생성한 mapper.xml과 연결 4. SelfIntroductionService 	mapper interface 와 연결 5. SelfIntroductionController 	각 기능마다 매개변수로 받아오는 값을 넣어주고 뷰로 연결되도록 구현

+++ linkJobs/src/main/java/com/gdu/linkJobs/controller/MemberCertificateController.java
@@ -22,13 +22,17 @@ public class MemberCertificateController {
 	//자격증 등록 폼
 	@GetMapping("/addMemberCertificate")
 	public String addMemberCertificate(HttpSession session) {
-		
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		return "member/addMemberCertificate";
 	}
 	//자격증 등록 액션
 	@PostMapping("/addMemberCertificate")
 	public String addMemberCertificate(MemberCertificate memberCertificate,HttpSession session) {
-		
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		memberCertificateService.addMemberCertificate(memberCertificate);
 		
 		return "rediect:/member/getMemberCertificateList";
@@ -38,6 +42,9 @@ public class MemberCertificateController {
 	@GetMapping("/modifyMemberCertificate")
 	public String modifyMemberCertificate(@RequestParam("certificateNo") int certificateNo,
 									Model model, MemberCertificate memberCertificate,HttpSession session) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		MemberCertificate MC= memberCertificateService.getMemberCertificateOne(certificateNo);
 		model.addAttribute("memberCertificate", MC);
 		return "member/modifyMemberCertificate";
@@ -45,6 +52,9 @@ public class MemberCertificateController {
 	//자격증 수정 액션
 	@PostMapping("/modifyMemberCertificate")
 	public String modifyMemberCertificate(MemberCertificate memberCertificate,HttpSession session) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		memberCertificateService.modifyMemberCertificate(memberCertificate);
 		return "rediect:/member/getMemberCertificateList";
 	}
@@ -52,14 +62,19 @@ public class MemberCertificateController {
 	//자격증 삭제
 	@GetMapping("/removeMemberCertificate")
 	public String removeMemberCertificate(@RequestParam("certificateNo") int certificateNo,HttpSession session) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		memberCertificateService.removeMemberCertificate(certificateNo);
 		return "rediect:/member/getMemberCertificateList";
 	}
 	
 	//자격증 회원별 리스트 출력 폼
 	@GetMapping("/getMemberCertificateList")
-	public String getMemberCertificateList(@RequestParam("memberId") String memberId,
-								Model model,HttpSession session) {
+	public String getMemberCertificateList(@RequestParam("memberId") String memberId, Model model,HttpSession session) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		List<MemberCertificate> certificateList=memberCertificateService.getMemberCertificateList(memberId);
 		model.addAttribute("certificateList", certificateList);
 		return "member/getMemberCertificateList";
